requirements.txt
amqp==5.1.1
asgiref==3.6.0
billiard==3.6.4.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.1.7
django-cors-headers==3.14.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
kombu==5.2.4
openpyxl==3.1.2
prompt-toolkit==3.0.38
pytz==2023.3
six==1.16.0
sqlparse==0.4.3
vine==5.0.0
wcwidth==0.2.6
